build: Move to standard Gradle project.
