+++ .gitignore
@@ -1,3 +1,8 @@
 /out
 /.idea
 /AutoCraft.iml
+# Ignore Gradle project-specific cache directory
+.gradle
+
+# Ignore Gradle build output directory
+build
